requirements.txt
PyYAML>=6.0
arxiv>=2.1.0
requests>=2.31.0